Fix ThreadIO dying when data is too large
Max data size in CachedData is not enough.
Indeed Memcached configuration is about the max item size which includes:
key, value and the item_struct sizes.
Checking the value size is not enough. ERR_2BIG can be triggered.

When EER_2BIG is triggered, the operation is marked as completed as soon as the server has answered with the error code.
The write operation is still in the write queue as there are bytes remaining to be written.
A NullPointerException is then throwned since the operation buffer is cleared when it is marked completed.
Finally the ThreadIO dies as it does not catch this type of exception.