Allow Optional Configuration of Max Clone Count
Allows the user to optionally change the max clone count from the
default of 100 to a value of their choosing via the
net.spy.maxCloneCount property. This enhancement is useful for
applications that have a large number of endpoints and want to use
the Redistribute FailureMode, but also want to limit the number of
cloning opportunities (as 100 is fairly high for a high throughput
usecase).