Use PropertyUtils to retrieve system properties.
The new PropertyUtils class wraps System.getProperty calls and handles SecurityExceptions. When a SecurityException occurs because the SecurityManager has denied access to the property read, null or the default value is returned. All code in spymemcached was already checking for null or setting a default value but it required grants to read the various properties. This change eliminates the need for those grants, making it easy to run spymemcached under a SecurityManager.